Squash the service controller and the service reload together
Having one single controller doing both the full reload and handling
service events prevents races, especially when handling delete events.
